massive code cleanup
