+++ clients/intro_course_developer_component/src/introCourse/pages/SeatAssignment/interfaces/DeveloperWithProfile.ts
@@ -1,0 +1,7 @@
+import { CoursePhaseParticipationWithStudent } from '@tumaet/prompt-shared-state'
+import { DeveloperProfile } from '../../../interfaces/DeveloperProfile'
+
+export type DeveloperWithProfile = {
+  participation: CoursePhaseParticipationWithStudent
+  profile: DeveloperProfile | undefined
+}
