+++ clients/intro_course_developer_component/src/introCourse/pages/SeatAssignment/utils/getTutorName.ts
@@ -1,6 +1,6 @@
 import { Tutor } from '../../../interfaces/Tutor'
 
-export const getTutorName = (tutorId: string, tutors: Tutor[]): string => {
+export const getTutorName = (tutorId: string | null, tutors: Tutor[]): string => {
   const tutor = tutors.find((t) => t.id === tutorId)
   return tutor ? `${tutor.firstName} ${tutor.lastName}` : 'Unknown Tutor'
 }
